feat: extract Client feature into ClientIrExtension
Introduce a generic IrExtension pattern so IrEmitter features can be
grouped and plugged in rather than baked into the emitter interface.

- Add IrExtension marker interface in ir.extensions.
- Add ClientIrExtension interface plus neutral helpers
  (convertEndpointClient, convertClient, buildClientServerInterfaces)
  in ir.extensions.ClientIrExtension.kt. Moved from IrConverter.kt
  and SharedBuilder.kt (renamed to AccessorStyle.kt).
- IrEmitter now exposes `val extensions: List<IrExtension>` and
  `inline fun extension<T>()` for typed lookup. emitClient and
  emitEndpointClient are no longer on the interface.
- Each language emitter (Java, Kotlin, Python, Rust, Scala,
  TypeScript) registers a per-language XxxClientIrExtension in its
  extensions list. Private client-only helpers moved into those
  extension classes (addIdentityTypeToCall, addSelfReceiverToClientFields,
  snakeCaseClientFunctions, flattenEndpointTypeRefs, buildClientParams).
- Java's transformTypeDescriptors moves to a standalone
  JavaTypeDescriptorTransform.kt so both the emitter and the client
  extension can use it.

All JVM tests pass; no behavioral changes to generated output.
