+++ src/compiler/emitters/python/src/commonMain/kotlin/community/flock/wirespec/emitters/python/PythonIrEmitter.kt
@@ -10,6 +10,7 @@ import community.flock.wirespec.ir.emit.SanitizationConfig
 import community.flock.wirespec.ir.emit.sanitizeFieldName
 import community.flock.wirespec.ir.emit.sanitizeNames
 import community.flock.wirespec.ir.emit.wrapWithModuleImport
+import community.flock.wirespec.ir.extensions.IrExtension
 import community.flock.wirespec.compiler.core.emit.Keywords
 import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToUpper
 import community.flock.wirespec.compiler.core.emit.PackageName
@@ -270,49 +271,13 @@ open class PythonIrEmitter(
         channel.convert()
             .sanitizeNames(sanitizationConfig)
 
-    override fun emitEndpointClient(endpoint: Endpoint): File {
-        val modelImports = endpoint.importReferences().distinctBy { it.value }
-            .map { Import("..model.${it.value}", LanguageType.Custom(it.value)) }
-        val endpointImport = Import("..endpoint.${endpoint.identifier.value}", LanguageType.Custom("*"))
-        val endpointName = endpoint.identifier.value
-
-        val file = super.emitEndpointClient(endpoint)
-            .sanitizeNames(sanitizationConfig)
-            .addSelfReceiverToClientFields()
-            .snakeCaseClientFunctions()
-            .flattenEndpointTypeRefs(endpointName)
-
-        val subPackageName = packageName + "client"
-        return File(
-            name = Name.of(subPackageName.toDir() + file.name.pascalCase()),
-            elements = buildImports("..wirespec") +
-                modelImports +
-                listOf(endpointImport) +
-                file.elements
-        )
-    }
-
-    override fun emitClient(endpoints: List<Endpoint>, logger: Logger): File {
-        val modelImports = endpoints.flatMap { it.importReferences() }.distinctBy { it.value }
-            .map { Import(".model.${it.value}", LanguageType.Custom(it.value)) }
-        val endpointImports = endpoints.map { Import(".endpoint.${it.identifier.value}", LanguageType.Custom("*")) }
-        val clientImports = endpoints.map { Import(".client.${it.identifier.value}Client", LanguageType.Custom("${it.identifier.value}Client")) }
-        val allImports = modelImports + endpointImports + clientImports
-        val endpointNames = endpoints.map { it.identifier.value }
-
-        val file = super.emitClient(endpoints, logger)
-            .sanitizeNames(sanitizationConfig)
-            .addSelfReceiverToClientFields()
-            .snakeCaseClientFunctions()
-            .let { f -> endpointNames.fold(f) { acc, name -> acc.flattenEndpointTypeRefs(name) } }
-
-        return File(
-            name = Name.of(packageName.toDir() + file.name.pascalCase()),
-            elements = buildImports(".wirespec") +
-                allImports +
-                file.elements
-        )
-    }
+    override val extensions: List<IrExtension> = listOf(
+        PythonClientIrExtension(
+            packageName = packageName,
+            sanitizationConfig = sanitizationConfig,
+            buildImports = ::buildImports,
+        ),
+    )
 
     private fun Identifier.sanitize(): String = value
         .split(".", " ")
@@ -363,64 +328,6 @@ open class PythonIrEmitter(
                     },
                 )
             } else iface
-        }
-    }
-
-    private fun <T : Element> T.flattenEndpointTypeRefs(endpointName: String): T = transform {
-        type { type, _ ->
-            if (type is LanguageType.Custom && type.name.startsWith("$endpointName.")) {
-                val suffix = type.name.removePrefix("$endpointName.")
-                if (suffix == "Call" || suffix == "Handler") type
-                else type.copy(name = suffix)
-            } else type
-        }
-    }
-
-    private fun <T : Element> T.addSelfReceiverToClientFields(): T {
-        val struct = (this as? File)?.findElement<Struct>()
-        val fieldNames = struct?.fields?.map { it.name.value() }?.toSet() ?: emptySet()
-        if (fieldNames.isEmpty()) return this
-
-        return transform {
-            statementAndExpression { stmt, tr ->
-                when (stmt) {
-                    is FieldCall -> {
-                        if (stmt.receiver == null && stmt.field.value() in fieldNames) {
-                            FieldCall(receiver = VariableReference(Name.of("self")), field = stmt.field)
-                        } else {
-                            FieldCall(
-                                receiver = stmt.receiver?.let { tr.transformExpression(it) },
-                                field = stmt.field,
-                            )
-                        }
-                    }
-                    else -> stmt.transformChildren(tr)
-                }
-            }
-        }
-    }
-
-    private fun <T : Element> T.snakeCaseClientFunctions(): T = transform {
-        matchingElements { func: LanguageFunction ->
-            func.copy(
-                name = Name.of(func.name.snakeCase()),
-                parameters = listOf(Parameter(Name.of("self"), LanguageType.Custom(""))) + func.parameters,
-            )
-        }
-        statementAndExpression { stmt, tr ->
-            when (stmt) {
-                is FunctionCall -> {
-                    val nameStr = stmt.name.value()
-                    val newName = if ("." in nameStr) stmt.name else Name.of(Name.of(nameStr).snakeCase())
-                    FunctionCall(
-                        name = newName,
-                        receiver = stmt.receiver?.let { tr.transformExpression(it) },
-                        arguments = stmt.arguments.mapValues { (_, v) -> tr.transformExpression(v) },
-                        isAwait = stmt.receiver != null,
-                    )
-                }
-                else -> stmt.transformChildren(tr)
-            }
         }
     }
