+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/emit/AccessorStyle.kt
@@ -1,0 +1,6 @@
+package community.flock.wirespec.ir.emit
+
+enum class AccessorStyle {
+    GETTER_METHODS,
+    PROPERTIES,
+}
