+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/emit/TestIrExtension.kt
@@ -1,0 +1,7 @@
+package community.flock.wirespec.ir.emit
+
+import community.flock.wirespec.ir.core.File
+
+fun interface TestIrExtension {
+    fun transformTestFile(file: File): File
+}
