+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/extensions/IrExtension.kt
@@ -1,0 +1,3 @@
+package community.flock.wirespec.ir.extensions
+
+interface IrExtension
